feat: Add LottoRule, Remove LottoPrice, Apply LottoNumber
