+++ src/main/java/lotto/LottoNumber.java
@@ -3,10 +3,7 @@ package lotto;
 import java.util.Objects;
 
 public class LottoNumber {
-    private static final int LOTTO_START_NUMBER = 1;
-    private static final int LOTTO_END_NUMBER = 45;
-
-    private int number;
+    private final int number;
 
     private LottoNumber(int number) {
         verify(number);
@@ -14,7 +11,7 @@ public class LottoNumber {
     }
 
     private void verify(int number) {
-        if (number < LOTTO_START_NUMBER || number > LOTTO_END_NUMBER) {
+        if (number < LottoRule.getLottoStartNumber() || number > LottoRule.getLottoEndNumber()) {
             throw new IllegalArgumentException(String
                     .format("%d는 1~45 숫자 사이의 번호가 아닙니다", number));
         }
