--- src/main/java/lotto/LottoPrice.java
@@ -1,9 +1,0 @@
-package lotto;
-
-public class LottoPrice {
-    private static final int PRICE = 1000;
-
-    public static int getPrice() {
-        return PRICE;
-    }
-}
