--- src/test/java/lotto/LottoPriceTest.java
@@ -1,15 +1,0 @@
-package lotto;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class LottoPriceTest {
-    @Test
-    @DisplayName("로또 가격 정책 가져오기")
-    void successByGetPrice() {
-        int price = LottoPrice.getPrice();
-        assertThat(price).isEqualTo(1000);
-    }
-}
